nlb alb loadbalancer 배포

+++ tf-aws-ekscube-01/loadbalancer-controller/create-policy-service-account.sh
@@ -4,12 +4,24 @@
 
 
 # IAM Policy download URL
-curl -O https://raw.githubusercontent.com/kubernetes-sigs/aws-load-balancer-controller/v2.13.3/docs/install/iam_policy.json
+if [ ! -f ./iam_policy.json ]; then
+  curl -O https://raw.githubusercontent.com/kubernetes-sigs/aws-load-balancer-controller/v2.13.3/docs/install/iam_policy.json
+fi
 
-# Create IAM policy
-aws iam create-policy \
-    --policy-name AWSLoadBalancerControllerIAMPolicy \
+POLICY_NAME=AWSLoadBalancerControllerIAMPolicy 
+
+POLICY_ARN=$(aws iam list-policies \
+  --scope Local \
+  --query "Policies[?PolicyName=='$POLICY_NAME'].Arn" \
+  --output text)
+
+if [ ! -n "$POLICY_ARN" ]; then
+  # Create IAM policy
+  aws iam create-policy \
+    --policy-name $POLICY_NAME \
     --policy-document file://iam_policy.json
+fi
+
 
 # Create Load Balancer Controller service account
 # cluser, attach-policy-arn, and region need to be updated
@@ -17,7 +29,8 @@ eksctl create iamserviceaccount \
     --cluster=eks-cicd-cluster \
     --namespace=kube-system \
     --name=aws-load-balancer-controller \
-    --attach-policy-arn=arn:aws:iam::533267419021:policy/AWSLoadBalancerControllerIAMPolicy \
+    --attach-policy-arn="arn:aws:iam::533267419021:policy/$POLICY_NAME" \
     --override-existing-serviceaccounts \
     --region ap-northeast-2 \
     --approve
+
